relocating ErrorBoundary
