+++ app/exercises/minesweeper/page.tsx
@@ -1,7 +1,7 @@
 "use client";
-import React, { useCallback, useEffect, useRef, useState } from "react";
+import React, { useEffect, useState } from "react";
 
-import ErrorBoundary from "./internals/ErrorBoundary";
+import ErrorBoundary from "../../lib/Components/ErrorBoundary";
 import { Levels } from "./internals/types";
 import { Board } from "./internals/styled/styled";
 import Tile from "./internals/Tile";
